eeprom: at24: fix 'assignment from incompatible pointer type' warning
The patch fixes this compiler's warning:

      CC      drivers/eeprom/at24.o
    drivers/eeprom/at24.c: In function ‘at24_probe’:
    drivers/eeprom/at24.c:453:21: warning: assignment from incompatible
    pointer type [-Wincompatible-pointer-types]
      at24->fops.protect = at24_cdev_protect,
                         ^

+++ drivers/eeprom/at24.c
@@ -352,7 +352,7 @@ static ssize_t at24_cdev_write(struct cdev *cdev, const void *buf, size_t count,
 	return at24_write(at24, buf, off, count);
 }
 
-static ssize_t at24_cdev_protect(struct cdev *cdev, size_t count, loff_t offset,
+static int at24_cdev_protect(struct cdev *cdev, size_t count, loff_t offset,
 		int prot)
 {
 	struct at24_data *at24 = cdev->priv;
